requirements.txt

Flask
gunicorn
sqlalchemy
python-dotenv
